Finalised logical device and graphic queue creation

+++ src/queue_families.rs
@@ -1,5 +1,5 @@
 pub struct QueueFamilyIndice {
-    pub graphics_family: Option<usize>,
+    pub graphics_family: Option<u32>,
 }
 
 impl QueueFamilyIndice {
